backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
sentence-transformers
youtube-transcript-api
yt-dlp
openai-whisper
torch
openai
pydantic
